requirements.txt
discord==2.3.2
qrcode==8.2
python-dotenv==1.1.1
Pillow